move owner-types to bases
