+++ backend/src/core/db_models/base.py
@@ -1,5 +1,8 @@
 from sqlalchemy import MetaData
+from sqlalchemy.dialects.postgresql import ENUM
 from sqlalchemy.orm import DeclarativeBase, registry
+
+from core.public_models.base import OwnerType
 
 DECLARED_METADATA = MetaData()
 DECLARED_REGISTRY = registry(metadata=DECLARED_METADATA)
@@ -14,3 +17,6 @@ class Base(DeclarativeBase):
 
     metadata = DECLARED_METADATA
     registry = DECLARED_REGISTRY
+
+
+DbOwnerType = ENUM(OwnerType)
