+++ backend/src/core/prompt_mgr/initial_prompts.py
@@ -3,7 +3,7 @@ from pathlib import Path
 
 import yaml
 
-from ..public_models import OwnerType
+from ..public_models.base import OwnerType
 from ..signals import db_predefined_data_handler
 from .util import add_chat_prompt
