+++ backend/src/core/prompt_mgr/prompt/query.py
@@ -7,7 +7,8 @@ from sqlalchemy.orm import aliased
 
 from ...db_models import DbAgentPrompt
 from ...providers.sql_database import DataDomain, get_sessionmaker
-from ...public_models import AgentPrompt, AgentPromptList, AgentPromptStatus, OwnerType, SortDirection
+from ...public_models import AgentPrompt, AgentPromptList, AgentPromptStatus, SortDirection
+from ...public_models.base import OwnerType
 from .stats import get_prompt_statistics
 
 logger = logging.getLogger(__name__)
