+++ backend/src/core/prompt_mgr/util.py
@@ -2,7 +2,8 @@ import textwrap
 from typing import Optional
 
 from ..agent.internal_models import AgentPromptName
-from ..public_models import AgentPromptStatus, OwnerType
+from ..public_models import AgentPromptStatus
+from ..public_models.base import OwnerType
 from .prompt.add import add_prompt
 from .prompt.query import list_prompts
